Documentation, Modules
Added documentation for front-end. Split the app into modules.
